"20 Question Clone"
After inputting a sentence, the program will try to figure out what the
user is describing. The tree.txt file is the format in which a tree can
be imported into the program. This contains the information that the
program can draw from in order to make a correct classification. The
program will always make a classification whether right or wrong so it
is not a true 20 Questions as it is limited to the character
information stored on the children of the tree imported. A much much
larger tree and could be more accurate, but the overall program works
correctly when staying within the bounds of the imported tree.

Some other options in the program are to edit the tree and